// commission/src/modules/grpc/proto/commission.pb.ts
export interface CommissionAnalysis {
  commissionTotalPercentage: number;
  commissionTotalFixed: number;
  paymentCommissionExtraFees: number;
  paymentCommission: number;
  paymentCommissionVat: number;
  nonPaymentCommission: number;
  nonPaymentCommissionVat: number;
  paymentCommissionWithVat: number;
  nonPaymentCommissionWithVat: number;
  penaltyCommission: number;
  penaltyCommissionVat: number;
  realEstateVat: number;
}

export interface CalculateCommissionSummaryResponse {
  commissionSummaries: BreakDownResponse[];
  productId: string;
}

export interface CalculateCommissionSummaryResponseForList {
  calculateCommissionSummaryResponses: CalculateCommissionSummaryResponse[];
}

export interface BreakDownResponse {
  withPromo: CommissionSummaryResponse | undefined;
  withoutPromo: CommissionSummaryResponse | undefined;
}

export interface CalculateCommissionSummaryRequest {
  commission: CommissionFilters | undefined;
  product: Product | undefined;
  calculationSettings: CalculationSettings | undefined;
  promoCode: PromoCode | undefined;
  allPayments: boolean;
  paymentModuleName: string;
  paymentOption: PaymentOption | undefined;
  reservation: Reservation | undefined;
  addonIds: string[];
  financingRequest: FinancingRequest | undefined;
}

export interface CalculateCommissionSummaryRequestForList {
  calculateCommissionSummaryRequests: CalculateCommissionSummaryRequest[];
}

export interface CreateCommissionSummaryRequest {
  commission: CommissionFilters | undefined;
  product: Product | undefined;
  calculationSettings: CalculationSettings | undefined;
  promoCode: PromoCode | undefined;
  order: Order | undefined;
  paymentModuleName: string;
  paymentOption: PaymentOption | undefined;
  reservation: Reservation | undefined;
  addOns: AddOn[];
  financingRequest: FinancingRequest | undefined;
}

export interface UpdateSellerCommissionRequest {
  product: Product | undefined;
  sellerNewCommission: number;
}

export interface ForceUpdateCommissionRequest {
  productId: string;
  orderId: string;
  grandTotal: number;
  payout: number;
  discount: number;
  buyerCommission: number;
  sellPrice: number;
}

export interface UpdateSellPriceRequest {
  product: Product | undefined;
}

export interface MigrateCommissionSummaryRequest {
  commission: CommissionFilters | undefined;
  product: Product | undefined;
  calculationSettings: MigrationCalculationSettings | undefined;
  promoCode: PromoCode | undefined;
  order: Order | undefined;
  paymentModuleName: string;
}

export interface Order {
  id: string;
}

export interface AddOn {
  id: string;
  addOnPrice: number;
}

export interface PaymentOption {
  id: string;
  paymentProvider: string;
  paymentCardType: string;
}

export interface CommissionFilters {
  userType: string;
  isBuyer: boolean;
}

export interface Product {
  id: string;
  sellPrice: number;
  priceRange: string;
  source: string;
  categoryId: string;
  modelId: string;
  varientId: string;
  conditionId: string;
}

export interface CalculationSettings {
  vatPercentage: number;
  applyDeliveryFeeSPPs: boolean;
  applyDeliveryFeeMPPs: boolean;
  applyDeliveryFee: boolean;
  deliveryFeeThreshold: number;
  deliveryFee: number;
  referralFixedAmount: number;
  applyReservation: boolean;
  applyFinancing: boolean;
}

export interface MigrationCalculationSettings {
  vatPercentage: number;
  applyDeliveryFeeSPPs: boolean;
  applyDeliveryFeeMPPs: boolean;
  applyDeliveryFee: boolean;
  deliveryFeeThreshold: number;
  deliveryFee: number;
  referralFixedAmount: number;
  buyerCommissionPercentage: number;
  sellerCommissionPercentage: number;
  priceQualityExtraCommission: number;
}

export interface Reservation {
  reservationAmount: number;
  reservationRemainingAmount: number;
}

export interface FinancingRequest {
  amount: number;
}

export interface PromoCode {
  promoLimit: number;
  type: string;
  generator: string;
  discount: number;
  percentage: number;
}

export interface CommissionSummaryResponse {
  id: string;
  commission: number;
  commissionVat: number;
  deliveryFee: number;
  deliveryFeeVat: number;
  totalVat: number;
  discount: number;
  grandTotal: number;
  commissionDiscount: number;
  sellPrice: number;
  commissionAnalysis: CommissionAnalysis | undefined;
  paymentId: string;
  reservation: Reservation | undefined;
  addOnsTotal: number;
  addOnsVat: number;
  addOnsGrandTotal: number;
  realEstateVat: number;
  financingRequest: FinancingRequest | undefined;
  paymentCardType?: string | undefined;
}

export interface ProductCommissionSummaryRequest {
  orderId: string;
  productId: string;
  isBuyer: boolean;
  isOriginalBreakDown: boolean;
}

export interface ProductCommissionSummaryResponse {
  id: string;
  commission: number;
  commissionVat: number;
  deliveryFee: number;
  deliveryFeeVat: number;
  totalVat: number;
  discount: number;
  grandTotal: number;
  commissionDiscount: number;
  productId: string;
  orderId: string;
  sellPrice: number;
  commissionAnalysis: CommissionAnalysis | undefined;
  reservation: Reservation | undefined;
  financingRequest: FinancingRequest | undefined;
}

export interface CalculateAddonSummaryRequest {
  productPrice: number;
  addonSummaryCalculateData: AddonSummaryCalculateData[];
}

export interface AddonSummaryCalculateData {
  priceType: string;
  addonPrice: number;
}

export interface CalculateAddonSummaryResponse {
  addOnsTotal: number;
  addOnsVat: number;
  addOnsGrandTotal: number;
}

export interface UpdateUsageCountRequest {
  promoCodeId: string;
  count: number;
}

export interface UpdateUsageCountResponse {
  ok: boolean;
}

export interface GetPromoDetailsRequest {
  filterField: string;
  filterFieldValue: string;
}

export interface PromoCodeScope {
  promoCodeScopeType: string;
  ids: string[];
}

export interface AvailablePayment {
  paymentProvider: string;
  paymentProviderType: string;
}

export interface DetailedPromoCode {
  promoLimit?: number | undefined;
  promoType: string;
  promoGenerator: string;
  discount?: number | undefined;
  percentage?: number | undefined;
  id: string;
  userType: string;
  status: string;
  code: string;
  isDefault: boolean;
  promoCodeScope: PromoCodeScope[];
  availablePayment: AvailablePayment[];
}

export interface GetDefaultPromoCodeRequest {
}

export interface GetFeedPromosRequest {
  feedIds: string[];
}

export interface GetFeedPromosResponse {
  DetailedPromoCode: DetailedPromoCode[];
}

export interface GetFeedPromoRequest {
  feedId: string;
}

export interface GetPromosByIdsRequest {
  ids: string[];
}

export interface GetPromosByIdsResponse {
  promos: DetailedPromoCode[];
}
// ...
export const COMMISSION_SERVICE_NAME = "CommissionService";